+++ library/src/main/jni/image/image.h
@@ -63,7 +63,7 @@ void decode(Stream* stream, bool partially, bool* animated, void** image);
 bool decode_info(Stream* stream, ImageInfo* info);
 
 bool decode_buffer(Stream* stream, bool clip, uint32_t x, uint32_t y, uint32_t width,
-    uint32_t height, uint8_t config, uint32_t ratio, BufferContainer* container);
+    uint32_t height, int32_t config, uint32_t ratio, BufferContainer* container);
 
 StaticImage* create(uint32_t width, uint32_t height, const uint8_t* data);
