+++ library/src/main/jni/image/image_jpeg.c
@@ -144,7 +144,7 @@ end:
 }
 
 bool jpeg_decode_buffer(Stream* stream, bool clip, uint32_t x, uint32_t y, uint32_t width,
-    uint32_t height, uint8_t config, uint32_t ratio, BufferContainer* container) {
+    uint32_t height, int32_t config, uint32_t ratio, BufferContainer* container) {
   struct jpeg_decompress_struct cinfo;
   struct my_error_mgr jerr;
   bool too_small;
