+++ library/src/main/jni/image/image_png.c
@@ -636,7 +636,7 @@ bool png_decode_info(Stream* stream, ImageInfo* info) {
 }
 
 bool png_decode_buffer(Stream* stream, bool clip, uint32_t x, uint32_t y, uint32_t width,
-    uint32_t height, uint8_t config, uint32_t ratio, BufferContainer* container) {
+    uint32_t height, int32_t config, uint32_t ratio, BufferContainer* container) {
   png_structp png_ptr = NULL;
   png_infop info_ptr = NULL;
   uint32_t i;
